feat: render text inserted from assistant
